refactor: move content addressed chunk functionality to cac
