+++ src/chunk/bmt.ts
@@ -1,6 +1,6 @@
 import { keccak256 } from 'js-sha3'
 import { BeeArgumentError } from '../utils/error'
-import { Bytes } from './bytes'
+import { Bytes } from '../utils/bytes'
 import { keccak256Hash } from './hash'
 
 const MAX_CHUNK_PAYLOAD_SIZE = 4096
