+++ src/chunk/soc.ts
@@ -1,45 +1,20 @@
-import { Bytes, FlexBytes, verifyBytes, verifyFlexBytes } from './bytes'
+import { Bytes, bytesAtOffset, bytesEqual, flexBytesAtOffset, verifyBytes } from '../utils/bytes'
 import { bmtHash } from './bmt'
 import { EthAddress, recoverAddress, sign, Signature, Signer } from './signer'
 import { keccak256Hash } from './hash'
-import { makeSpan } from './span'
+import { SPAN_SIZE } from './span'
 import { serializeBytes } from './serialize'
 import { BeeError } from '../utils/error'
 import { BrandedType } from '../types'
+import { Chunk, ChunkAddress, MAX_PAYLOAD_SIZE, MIN_PAYLOAD_SIZE, verifyChunk } from './cac'
 
-const MIN_PAYLOAD_SIZE = 1
-const MAX_PAYLOAD_SIZE = 4096
-
-const SPAN_SIZE = 8
 const IDENTIFIER_SIZE = 32
 const SIGNATURE_SIZE = 65
-
-const CAC_SPAN_OFFSET = 0
-const CAC_PAYLOAD_OFFSET = CAC_SPAN_OFFSET + SPAN_SIZE
 
 const SOC_IDENTIFIER_OFFSET = 0
 const SOC_SIGNATURE_OFFSET = SOC_IDENTIFIER_OFFSET + IDENTIFIER_SIZE
 const SOC_SPAN_OFFSET = SOC_SIGNATURE_OFFSET + SIGNATURE_SIZE
 const SOC_PAYLOAD_OFFSET = SOC_SPAN_OFFSET + SPAN_SIZE
-
-type ChunkAddress = Bytes<32>
-
-/**
- * General chunk interface for Swarm
- *
- * It stores the serialized data and provides functions to access
- * the fields of a chunk.
- *
- * It also provides an address function to calculate the address of
- * the chunk that is required for the Chunk API.
- */
-export interface Chunk {
-  readonly data: Uint8Array
-  span(): Bytes<8>
-  payload(): FlexBytes<1, 4096>
-
-  address(): ChunkAddress
-}
 
 export type Identifier = Bytes<32>
 
@@ -56,67 +31,7 @@ export interface SingleOwnerChunk extends Chunk {
   signature: () => Signature
 }
 
-type ValidChunkData = BrandedType<Uint8Array, 'ValidChunkData'>
 type ValidSingleOwnerChunkData = BrandedType<Uint8Array, 'ValidSingleOwnerChunkData'>
-
-type SpanReference = Bytes<32> | Bytes<64>
-
-/**
- * Creates a content addressed chunk and verifies the payload size.
- *
- * @param payloadBytes the data to be stored in the chunk
- */
-export function makeContentAddressedChunk(payloadBytes: Uint8Array): Chunk {
-  const span = makeSpan(payloadBytes.length)
-  const payload = verifyFlexBytes(MIN_PAYLOAD_SIZE, MAX_PAYLOAD_SIZE, payloadBytes)
-  const data = serializeBytes(span, payload) as ValidChunkData
-  const address = () => bmtHash(data)
-
-  return makeChunk(data, address)
-}
-
-function makeChunk(data: ValidChunkData, address: () => ChunkAddress): Chunk {
-  const span = () => bytesAtOffset(CAC_SPAN_OFFSET, SPAN_SIZE, data)
-  const payload = () => flexBytesAtOffset(CAC_PAYLOAD_OFFSET, MIN_PAYLOAD_SIZE, MAX_PAYLOAD_SIZE, data)
-
-  return {
-    data,
-    span,
-    payload,
-    address,
-  }
-}
-
-/**
- * Type guard for valid content addressed chunk data
- *
- * @param data          The chunk data
- * @param chunkAddress  The address of the chunk
- */
-export function isValidChunkData(data: Uint8Array, chunkAddress: ChunkAddress): data is ValidChunkData {
-  const address = bmtHash(data)
-
-  return bytesEqual(address, chunkAddress)
-}
-
-/**
- * Verifies if a chunk is a valid content addressed chunk
- *
- * @param data          The chunk data
- * @param chunkAddress  The address of the chunk
- *
- * @returns a valid content addressed chunk or throws error
- */
-export function verifyChunk(data: Uint8Array, chunkAddress: ChunkAddress): Chunk {
-  if (isValidChunkData(data, chunkAddress)) {
-    return makeChunk(data, () => chunkAddress)
-  }
-  throw new BeeError('verifyChunk')
-}
-
-function bytesEqual(a: Uint8Array, b: Uint8Array): boolean {
-  return a.length === b.length && a.every((value, index) => value === b[index])
-}
 
 /**
  * Type guard for valid single owner chunk data
@@ -153,7 +68,7 @@ export function isValidSingleOwnerChunkData(
  */
 export function verifySingleOwnerChunk(data: Uint8Array, address: ChunkAddress): SingleOwnerChunk {
   if (isValidSingleOwnerChunkData(data, address)) {
-    return makeSingleOwnerChunkFromData(data, () => address)
+    return makeSingleOwnerChunkFromData(data, address)
   }
   throw new BeeError('verifySingleOwnerChunk')
 }
@@ -162,24 +77,12 @@ function verifyBytesAtOffset<Length extends number>(offset: number, length: Leng
   return verifyBytes(length, bytesAtOffset(offset, length, data))
 }
 
-function bytesAtOffset<Length extends number>(offset: number, length: Length, data: Uint8Array): Bytes<Length> {
-  return data.slice(offset, offset + length) as Bytes<Length>
-}
-
-function flexBytesAtOffset<Min extends number, Max extends number>(
-  offset: number,
-  _min: Min,
-  _max: Max,
-  data: Uint8Array,
-): FlexBytes<Min, Max> {
-  return data.slice(offset) as FlexBytes<Min, Max>
-}
-
-function makeSingleOwnerChunkFromData(data: ValidSingleOwnerChunkData, address: () => ChunkAddress): SingleOwnerChunk {
+function makeSingleOwnerChunkFromData(data: ValidSingleOwnerChunkData, chunkAddress: ChunkAddress): SingleOwnerChunk {
   const identifier = () => bytesAtOffset(SOC_IDENTIFIER_OFFSET, IDENTIFIER_SIZE, data)
   const signature = () => bytesAtOffset(SOC_SIGNATURE_OFFSET, SIGNATURE_SIZE, data)
   const span = () => bytesAtOffset(SOC_SPAN_OFFSET, SPAN_SIZE, data)
   const payload = () => flexBytesAtOffset(SOC_PAYLOAD_OFFSET, MIN_PAYLOAD_SIZE, MAX_PAYLOAD_SIZE, data)
+  const address = () => chunkAddress
 
   return {
     data,
@@ -213,7 +116,7 @@ export async function makeSingleOwnerChunk(
   const digest = keccak256Hash(identifier, chunkAddress)
   const signature = await sign(digest, signer)
   const data = serializeBytes(identifier, signature, chunk.span(), chunk.payload()) as ValidSingleOwnerChunkData
-  const address = () => keccak256Hash(identifier, signer.address)
+  const address = keccak256Hash(identifier, signer.address)
 
   return makeSingleOwnerChunkFromData(data, address)
 }
